Allow selecting collapsed JSON text (part 2)
This cancels click events on collapsed JSON trees if the document has
some selected text.

Since the click handler is inside `react-json-tree` rather than code we
control, this is done using a document-level event handler with
`useCapture` set to true so that we can intercept the click event before
it fires.

See also:

- http://stackoverflow.com/a/4616720
- https://github.com/Automattic/rest-api-console2/pull/23