Remove `lodash` as a project dependency
Our use of the `babel-lodash-plugin` has made it hard to update our
build tool versions. It has also tied us to that plugin lest we
dramatically increase our build size.

In this patch we're replacing all uses of `lodash` functions with
alternatives, simplified copies of the `lodash` functions, or
with direct expressions equivalent to the `lodash` variants.