Remove unintentional optional chain introduced in #91
In #91 we accidentally introduced an optional chain into the source code
via accepting a change proposed during PR review through the Github interface.

It appears that the version of the project build and deployed did not include
the optional chain because it was built without pulling the updated changes
from the PR.

In this patch we're removing that and thankfully it wasn't necessary because
the optionality was already guarded against in the surrounding block.