More-completely parse endpoint PCRE pattern for better parameter discovery. (#102)
Replaces #101

In #101 we identified a situation in which unexpected input from the API discovery
crashed the API console. The situation was that the `type` or acceptible PCRE pattern
for a parameter involved a subgroup. Unfortunately we have been expecting no subgroups
to exist and therefore assumed we could infer parameter groups by looking for the first
closing parenthesis after the opening. See the example: a picture shows better.

    /wpcom/v2/jetpack-boost-proxy/(?P<version>(?:v\\d+/)?)metrics
                                  ^^^^^^^^^^^^|^^^^^^^^^^|
                    pattern group /           |          \- should end here
                                              |
                                              \- previously ended here

In this patch we're changing the parsing method to account for balanced parentheses
inside the parameter descriptions. This isn't a complete solution because it's
possible for unescaped parenthesis to exist within character groups of a PCRE pattern.

Initially after improving the parser the display of the parameters in the endpoint
bar was messed up because of the similarly naive approach for detecting parameters:
that they start with `$` and end with a `/`. In the example above we can see that
some parameters are given in terms of _part_ of a URL path part.

The display logic has been updated to search not for those `$`-prefixed segments
but rather to search directly for the `$`-prefix parameter names. This simplifies
the splitting logic for the `<LookupContainer>`.